[AS-606] layers clarification in progress
