--- src/config/index.ts
@@ -1,7 +1,0 @@
-
-
-export enum WalletClientModes {
-    LOCAL = "local",
-    MPC = "mpc",
-}
-
