--- src/domain/Signer/index.ts
@@ -1,15 +1,0 @@
-import Wallet from "../aggregates/Wallet";
-
-export interface SigningRequest {
-    wallet: Wallet;
-    data: any;
-}
-
-export interface SignedResult {
-    data: any;
-    deployer: string;
-    signature: string;
-    sigAlgorithm: string;
-}
-
-export type PasswordProvider = () => Promise<string>;
