+++ src/infrastructure/configuration/index.ts
@@ -1,4 +1,4 @@
-import { ResubmitConfig } from "@services/Resubmit/types";
+import { ResubmitConfig } from "../../application/services/Resubmit/types";
 export const DEFAULT_AXIOS_TIMEOUT_MS: number = 30000;
 export const DEFAULT_PHLO_LIMIT: number = 500000;
 export const DEFAULT_RESUBMIT_CONFIG: ResubmitConfig = {
