+++ src/services/KeyDerivation/index.ts
@@ -1,6 +1,6 @@
 import ECC from "./eccAdapter";
 import MnemonicService from "@services/Mnemonic";
-import { ASI_COIN_TYPE, DEFAULT_BIP_44_PATH_OPTIONS } from "@utils/constants";
+import { ASI_COIN_TYPE, DEFAULT_BIP_44_PATH_OPTIONS } from "@domain/constants";
 import { BIP32Factory, type BIP32Interface } from "bip32";
 import { setupBufferPolyfill } from "../../infrastructure/misc/polyfills";
 import { mnemonicToSeed } from "bip39";
