+++ src/services/index.ts
@@ -8,8 +8,8 @@ export * from "./Wallets";
 export * from "./Mnemonic";
 export * from "./KeysManager";
 export * from "./KeyDerivation";
-export * from "./Resubmit";    
-export * from "./Signer";
+export * from "../application/services/Resubmit";    
+export * from "../infrastructure/adapters/Signer";
 export * from "../application/services/AssetsService";
 export * from "../application/services/Fee";
 export * from "../application/services/TxHistory";
@@ -22,7 +22,7 @@ export { default as WalletsService } from "./Wallets";
 export { default as CryptoService } from "../infrastructure/adapters/Crypto";
 
 
-export { default as DeployResubmitter } from "./Resubmit";
-export { default as SignerService } from "./Signer";
+export { default as DeployResubmitter } from "../application/services/Resubmit";
+export { default as SignerService } from "../infrastructure/adapters/Signer";
 export { default as AssetsService } from "../application/services/AssetsService";
-export { default as FundsReservationService } from "./FundsReservation";
+export { default as FundsReservationService } from "../application/services/FundsReservation";
