+++ tests/reservation/reservation-functionality.test.ts
@@ -1,6 +1,6 @@
 import assert from "node:assert";
 import ReservationRecord from "../../src/domain/aggregates/Reservation";
-import FundsReservationService from "@services/FundsReservation";
+import FundsReservationService from "../../src/application/services/FundsReservation";
 import { describe, it, beforeEach, afterEach } from "node:test";
 import { ReservationStatus } from "../../src/domain/aggregates/Reservation/types";
 import { Address } from "../../src/domain/aggregates/Wallet";
